@@ -1,10 +1,10 @@
 name: drop_cap_text
 description: This Flutter plugin aims to achive drop cap in text capitalization.
-version: 1.1.3
+version: 2.0.0
 homepage: https://github.com/mtiziano/drop_cap_text
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -18,7 +18,6 @@ dev_dependencies:
 # following page: https://www.dartlang.org/tools/pub/pubspec
 
 # The following section is specific to Flutter.
-flutter:
 
   # To add assets to your package, add an assets section, like this:
   # assets:
